third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-tree-styles-with-apply-author-styles-expected.txt

Test for issue 247275: shadow tree styles selects shadow host when a shadow root has apply-author-styles true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test when one shadow root has apply-author-styles true
PASS borderColorOf('host/target') is "rgb(0, 128, 0)"
Test when an inert shadow root has apply-author-styles true
PASS borderColorOf('host/target') is "rgb(0, 0, 0)"
Test when one of shadow roots which are not out of compositing has apply-author-styles true
PASS borderColorOf('host/target') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
